added insertType to batch
