+++ packages/destination-actions/src/destinations/hubspot/index.ts
@@ -48,7 +48,7 @@ const destination: DestinationDefinition<Settings> = {
     return {
       skipResponseCloning: true,
       headers: {
-        authorization: `Bearer CMbwz8XzMRIVAAEBQAAA4QIAAAC4AAAAAAAQAAACGJr1gwsgosxAKJf4QTIUUPV2hOYBhrsIs_ZBe7B7R-f6C4Q6QAAAAEEAAAAAwAcAAAAAAAAAgAAAAAAAAAAMACAAjgA-AOABAAAAAAAA_GcAAADwAwAAAAAAAABAAAAAAAAEAAxCFAHqDXcfVpxf0uNZeQt_cq16k54ASgNuYTFSAFoAYAA`
+        authorization: `Bearer CO_SvsfzMRIVAAEBQAAA4QIAAAC4AAAAAAAQAAACGJr1gwsgosxAKJf4QTIUwhE7qnzEH-ZdMp--cd-UsyjSMUs6QAAAAEEAAAAAwAcAAAAAAAAAgAAAAAAAAAAMACAAjgA-AOABAAAAAAAA_GcAAADwAwAAAAAAAABAAAAAAAAEAAxCFJI6wfw2KPKu7kkdLVOMYvlGwPPASgNuYTFSAFoAYAA`
       }
     }
   },
