+++ packages/destination-actions/src/destinations/hubspot/upsertObject/constants.ts
@@ -1,1 +1,7 @@
 export const BATCH_SIZE = 2
+
+export const INSERT_TYPES = [    
+    {label: 'Create', value: 'create'},
+    {label: 'Update', value: 'update'},
+    {label: 'Upsert', value: 'upsert'}
+]
